+++ src/main.js
@@ -12,7 +12,7 @@ import 'vue-awesome/icons/arrow-right'
 import 'vue-awesome/icons/check'
 import 'vue-awesome/icons/chevron-left'
 import 'vue-awesome/icons/'
-import ForkeMeOnGithub from 'fork-me-on-github-vue'
+import ForkMeOnGithub from 'fork-me-on-github-vue'
 import BootstrapVue from 'bootstrap-vue'
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
@@ -43,7 +43,7 @@ const Defiant = require('./lib/defiant') // eslint-disable-line
 
 Vue.use(VTooltip)
 Vue.use(VueYouTubeEmbed)
-Vue.use(ForkeMeOnGithub)
+Vue.use(ForkMeOnGithub)
 Vue.use(BootstrapVue)
 Vue.component('star-rating', StarRating)
 Vue.component('img-left', ImageLeft)